Fixed UF-78: Resource leak in AbstractMultiPartWorkbenchPanelPresenter (was BaseMultiPartWorkbenchPanelPresenter)
The mocking in these tests is a bit over-the-top. It would be nice to clean it up some time.

Also, the tests don't cover all possible context resolution permutations. There are a lot of them. The tests mostly
focus on the code paths that ensure the resource leak test is doing something valid.