Moved responsibility for panel creation & composition away from Workbench and into PanelManager
This commit also simplifies the startup procedure for the workbench UI as well as some of the
resize logic (but there is more to do; see UF-112)

There is also a likely DnD resource leak in this change. We need a cleanup contract between PanelManager
and PanelPresenter so the presenter knows when to remove listeners from its PanelView.