Added new integration test to ensure MultiList panels are sized properly
This is a regression test for a fix that was committed on Friday.

This commit also introduces a new factory method for DefaultPlaceRequest
that can parse requests in the form PlaceId?param1=val1&param2=val2 ...
This was lifted from the annotation processing logic for @WorkbenchPanel.
It might come in useful for more stuff in the future.