Added an onResize() call at the end of perspective switching.
This should not be necessary for well-behaved apps, because the CSS sizing of all panels
should sort itself out without any help from the application code. But some apps choose to
set explicit sizes on widgets when receiving an onResize, and this sometimes ends up with
a 0px explicit size because the CSS updates sometimes haven't happened yet when the perspective
is being built. (We have found this is more common in Chrome than in FireFox).