Fix Duplicated PathPlaceRequest Activity
With the new mechanism to lookup PathPlaceRequest
identifier, I've introduced a bug.
In order to get the bean activity I've called
bean.getInstance() on activity bean.

The problem is that the secure method in ActivityManagerImpl
makes the same call and creates a new bean, making activity
beans bean created twice generating a activity leak and lock issues.